Fixed various clang compiler warnings

+++ benchmark/dot_product_BENCHMARK.cc
@@ -156,7 +156,7 @@ class DenseVectorDotProductSPUPlot :
             int temp(Configuration::instance()->get_value("cell::dot_product_dense_dense_float", 4));
             int temp2(Configuration::instance()->get_value("cell::dot_product_dense_dense_double", 4));
 
-            int max_spu(6);
+            unsigned long max_spu(6);
 
             for (unsigned long j(1) ; j <= max_spu ; ++j)
             {
